Re-extract progress, don't include number in result
