+++ controllers/apply/form-router-next/lib/enrich-form.js
@@ -2,67 +2,10 @@
 const cloneDeep = require('lodash/cloneDeep');
 const find = require('lodash/find');
 const flatMap = require('lodash/flatMap');
-const flatMapDeep = require('lodash/flatMapDeep');
-const get = require('lodash/get');
-const includes = require('lodash/includes');
-const isEmpty = require('lodash/isEmpty');
-const pick = require('lodash/pick');
 const reject = require('lodash/reject');
 
 const { formatterFor } = require('./formatters');
-const validateForm = require('./validate-form');
-
-/**
- * Calculate form progress
- * Validates the form and returns a status
- * for each section and the form as a whole.
- */
-function progressSummary(form, data) {
-    const FORM_STATES = {
-        empty: 'empty',
-        incomplete: 'incomplete',
-        complete: 'complete'
-    };
-
-    function determineStatus(value, errors = []) {
-        if (isEmpty(value)) {
-            return FORM_STATES.empty;
-        } else if (errors.length > 0) {
-            return FORM_STATES.incomplete;
-        } else {
-            return FORM_STATES.complete;
-        }
-    }
-
-    const validationResult = validateForm(form, data);
-
-    const errors = get(validationResult, 'error.details', []);
-
-    const allStatus = determineStatus(validationResult.value, errors);
-
-    return {
-        isComplete: allStatus === FORM_STATES.complete,
-        all: determineStatus(validationResult.value, errors),
-        sections: form.sections.map((section, idx) => {
-            const fieldNames = flatMapDeep(section.steps, step => {
-                return step.fieldsets.map(fieldset =>
-                    fieldset.fields.map(field => field.name)
-                );
-            });
-
-            const dataForSection = pick(validationResult.value, fieldNames);
-            const errorsForSection = errors.filter(detail =>
-                includes(fieldNames, detail.path[0])
-            );
-
-            return {
-                slug: section.slug,
-                label: `${idx + 1}: ${section.shortTitle || section.title}`,
-                status: determineStatus(dataForSection, errorsForSection)
-            };
-        })
-    };
-}
+const progress = require('./progress');
 
 /**
  * Enriches a form object
@@ -73,7 +16,7 @@ function progressSummary(form, data) {
 module.exports = function enrichForm(baseForm, data = {}) {
     const clonedForm = cloneDeep(baseForm);
 
-    const progress = progressSummary(clonedForm, data);
+    const progressSummary = progress(clonedForm, data);
 
     function enrichField(field) {
         // Assign value to field if present
@@ -112,7 +55,7 @@ module.exports = function enrichForm(baseForm, data = {}) {
     }
 
     clonedForm.sections = clonedForm.sections.map(section => {
-        section.progress = progress.sections.find(
+        section.progress = progressSummary.sections.find(
             progressSection => progressSection.slug === section.slug
         );
 
@@ -127,7 +70,7 @@ module.exports = function enrichForm(baseForm, data = {}) {
         clonedForm.termsFields = clonedForm.termsFields.map(enrichField);
     }
 
-    clonedForm.progress = progress;
+    clonedForm.progress = progressSummary;
 
     clonedForm.fullSummary = function() {
         const steps = flatMap(clonedForm.sections, 'steps');
