+++ controllers/apply/form-router-next/views/dashboard.njk
@@ -66,7 +66,7 @@
                 {% for section in application.progress.sections %}
                     <li class="section-progress-mini section-progress-mini--{{ section.status }}">
                         <span class="section-progress-mini__title">
-                            {{ section.label }}
+                            {{ loop.index }}: {{ section.label }}
                         </span>
                         <span class="section-progress-mini__marker">
                             {% if section.status === 'complete' %}
